Don't try to call the API to delete files when we are not logged in

+++ lib/src/app_logic.dart
@@ -1109,6 +1109,11 @@ class AppLogic {
 
   Future<bool> _deleteFile(File file) async {
     await _db.deleteFile(file);
+
+    if (appState.accountDetails.value == null) {
+      return true;
+    }
+
     final success = await _api.deleteFile(file.id);
 
     if (!success) {
